feat: общий модуль расчёта цены
